# packages/api-client-go/api/openapi.yaml
openapi: 3.0.0
info:
  contact: {}
  description: Snapflow AI platform API Docs
  title: Snapflow
  version: "1.0"
servers:
- url: http://localhost:3000
paths:
...
  /preview/{sandboxId}/public:
    get:
      operationId: isSandboxPublic
      parameters:
      - description: ID of the sandbox
        explode: false
        in: path
        name: sandboxId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                type: boolean
          description: Public status of the sandbox
      summary: Check if sandbox is public
      tags:
      - preview
...
components:
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      description: API Key access
      scheme: bearer
      type: http